Separate Bandcamp endpoint requests from main logic
